+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 0.0.4
 
+- Added CellId.rangeMin & CellId.rangeMax (thanks to @mast)
 
 ## Version 0.0.3